+++ typescript/packages/subsurface-viewer/src/storybook/layers/WellsLayer.stories.tsx
@@ -39,6 +39,7 @@ import {
     DEFAULT_TOLERANCE,
 } from "../../layers/wells/utils/spline";
 import {
+    LABEL_MERGE_RADIUS_ARGTYPES,
     LABEL_ORIENTATION_ARGTYPES,
     LABEL_POSITION_ARGTYPES,
     LABEL_SIZE_ARGTYPES,
@@ -696,6 +697,9 @@ export const WellLabelStyle: StoryObj<
         getBorderWidth: 1,
         bgColor: "white",
         getPositionAlongPath: 0,
+        mergeLabels: true,
+        mergeRadius: 100,
+        autoPosition: true,
     },
 
     render: (args) => {
@@ -709,26 +713,34 @@ export const WellLabelStyle: StoryObj<
             orientation: args.orientation,
             background: args.background,
             visible: true,
+            autoPosition: args.autoPosition,
+
+            // MergedTextLayer options
+            mergeLabels: args.mergeLabels,
+            mergeRadius: args.mergeRadius,
+
+            // Disable transitions as they are too slow for this story, on the test runner
+            transitions: {},
         };
 
-        const wellLayer = new WellsLayer({
+        const wellLayerProps = {
             data: "./gullfaks.json",
             wellHeadStyle: { size: 5 },
             ZIncreasingDownwards: false,
             wellLabel: wellLabelProps,
-            id: "wells",
+        };
+
+        const wellLayer2d = new WellsLayer({
+            ...wellLayerProps,
+            id: "wells-2d",
         });
 
-        const axes2d = new Axes2DLayer({
-            id: "axes-2d",
+        const wellLayer3d = new WellsLayer({
+            ...wellLayerProps,
+            id: "wells-3d",
         });
 
-        const axes3d = new AxesLayer({
-            id: "axes-3d",
-            bounds: [450000, 6781000, 0, 464000, 6791000, 3500],
-        });
-
-        const layers = [wellLayer, axes3d, axes2d];
+        const layers = [wellLayer2d, wellLayer3d];
 
         // Viewport is reset if identity of `views` object changes, so we need to memoize it.
         // eslint-disable-next-line react-hooks/rules-of-hooks
@@ -738,12 +750,12 @@ export const WellLabelStyle: StoryObj<
                 viewports: [
                     {
                         id: "viewport1",
-                        layerIds: ["wells", "axes-3d"],
+                        layerIds: ["wells-3d"],
                         show3D: true,
                     },
                     {
                         id: "viewport2",
-                        layerIds: ["wells", "axes-2d"],
+                        layerIds: ["wells-2d"],
                         show3D: false,
                     },
                 ],
@@ -764,6 +776,7 @@ export const WellLabelStyle: StoryObj<
         ...LABEL_POSITION_ARGTYPES,
         ...LABEL_SIZE_ARGTYPES,
         ...LABEL_ORIENTATION_ARGTYPES,
+        ...LABEL_MERGE_RADIUS_ARGTYPES,
     },
 };